[dyad] Refactored gallery upload logic into new utility file - wrote 2 file(s)
